+++ examples/bignumbers.js
@@ -11,8 +11,8 @@ math.config({
 });
 
 console.log('round-off errors with numbers');
-print(math.add(0.1, 0.2));    // Number, 0.30000000000000004
-print(math.divide(0.3, 0.2)); // Number, 1.4999999999999998
+print(math.add(0.1, 0.2));    // number, 0.30000000000000004
+print(math.divide(0.3, 0.2)); // number, 1.4999999999999998
 console.log();
 
 console.log('no round-off errors with BigNumbers');