+++ test/expression/operators.test.js
@@ -8,8 +8,6 @@ var ConstantNode = math.expression.node.ConstantNode;
 var Node = math.expression.node.Node;
 var ParenthesisNode = math.expression.node.ParenthesisNode;
 
-var config = {parenthesis: 'keep'};
-
 describe('operators', function () {
   it('should return the precedence of a node', function () {
     var a = new ConstantNode(1);
@@ -18,14 +16,14 @@ describe('operators', function () {
     var n1 = new AssignmentNode('a', a);
     var n2 = new OperatorNode('or', 'or', [a, b]);
 
-    assert.equal(operators.getPrecedence(n1, config), 0);
-    assert.equal(operators.getPrecedence(n2, config), 2);
+    assert.equal(operators.getPrecedence(n1, 'keep'), 0);
+    assert.equal(operators.getPrecedence(n2, 'keep'), 2);
   });
 
   it('should return null if precedence is not defined for a node', function () {
     var n = new Node();
 
-    assert.equal(operators.getPrecedence(n, config), null);
+    assert.equal(operators.getPrecedence(n, 'keep'), null);
   });
 
   it ('should return the precedence of a ParenthesisNode', function () {
@@ -35,9 +33,9 @@ describe('operators', function () {
 
     var p = new ParenthesisNode(op);
 
-    assert.equal(operators.getPrecedence(p, {parenthesis: 'all'}), operators.getPrecedence(op, config));
-    assert.equal(operators.getPrecedence(p, {parenthesis: 'auto'}), operators.getPrecedence(op, config));
-    assert.equal(operators.getPrecedence(p, config), null);
+    assert.equal(operators.getPrecedence(p, 'all'), operators.getPrecedence(op, 'all'));
+    assert.equal(operators.getPrecedence(p, 'auto'), operators.getPrecedence(op, 'all'));
+    assert.equal(operators.getPrecedence(p, 'keep'), null);
   });
 
   it('should return the associativity of a node', function () {
@@ -48,10 +46,10 @@ describe('operators', function () {
     var n3 = new OperatorNode('-', 'unaryMinus', [a]);
     var n4 = new OperatorNode('!', 'factorial', [a]);
 
-    assert.equal(operators.getAssociativity(n1, config), 'left');
-    assert.equal(operators.getAssociativity(n2, config), 'right');
-    assert.equal(operators.getAssociativity(n3, config), 'right');
-    assert.equal(operators.getAssociativity(n4, config), 'left');
+    assert.equal(operators.getAssociativity(n1, 'keep'), 'left');
+    assert.equal(operators.getAssociativity(n2, 'keep'), 'right');
+    assert.equal(operators.getAssociativity(n3, 'keep'), 'right');
+    assert.equal(operators.getAssociativity(n4, 'keep'), 'left');
   });
 
   it ('should return the associativity of a ParenthesisNode', function () {
@@ -61,9 +59,9 @@ describe('operators', function () {
 
     var p = new ParenthesisNode(op);
 
-    assert.equal(operators.getAssociativity(p, {parenthesis: 'all'}), operators.getAssociativity(op, config));
-    assert.equal(operators.getAssociativity(p, {parenthesis: 'auto'}), operators.getAssociativity(op, config));
-    assert.equal(operators.getAssociativity(p, config), null);
+    assert.equal(operators.getAssociativity(p, 'all'), operators.getAssociativity(op, 'keep'));
+    assert.equal(operators.getAssociativity(p, 'auto'), operators.getAssociativity(op, 'keep'));
+    assert.equal(operators.getAssociativity(p, 'keep'), null);
   });
 
   it('should return null if associativity is not defined for a node', function () {
@@ -72,8 +70,8 @@ describe('operators', function () {
     var n1 = new Node();
     var n2 = new AssignmentNode('a', a);
 
-    assert.equal(operators.getAssociativity(n1, config), null);
-    assert.equal(operators.getAssociativity(n2, config), null);
+    assert.equal(operators.getAssociativity(n1, 'keep'), null);
+    assert.equal(operators.getAssociativity(n2, 'keep'), null);
   });
 
   it('should return if a Node is associative with another Node', function () {
@@ -82,10 +80,10 @@ describe('operators', function () {
     var n1 = new OperatorNode('+', 'add', [a, a]);
     var n2 = new OperatorNode('-', 'subtract', [a, a]);
 
-    assert.equal(operators.isAssociativeWith(n1, n1, config), true);
-    assert.equal(operators.isAssociativeWith(n1, n2, config), true);
-    assert.equal(operators.isAssociativeWith(n2, n2, config), false);
-    assert.equal(operators.isAssociativeWith(n2, n1, config), false);
+    assert.equal(operators.isAssociativeWith(n1, n1, 'keep'), true);
+    assert.equal(operators.isAssociativeWith(n1, n2, 'keep'), true);
+    assert.equal(operators.isAssociativeWith(n2, n2, 'keep'), false);
+    assert.equal(operators.isAssociativeWith(n2, n1, 'keep'), false);
   });
 
   it('should return null if the associativity between two Nodes is not defined', function () {
@@ -94,10 +92,10 @@ describe('operators', function () {
     var n1 = new Node();
     var n2 = new AssignmentNode('a', a);
 
-    assert.equal(operators.isAssociativeWith(n1, n1, config), null);
-    assert.equal(operators.isAssociativeWith(n1, n2, config), null);
-    assert.equal(operators.isAssociativeWith(n2, n2, config), null);
-    assert.equal(operators.isAssociativeWith(n2, n1, config), null);
+    assert.equal(operators.isAssociativeWith(n1, n1, 'keep'), null);
+    assert.equal(operators.isAssociativeWith(n1, n2, 'keep'), null);
+    assert.equal(operators.isAssociativeWith(n2, n2, 'keep'), null);
+    assert.equal(operators.isAssociativeWith(n2, n1, 'keep'), null);
   });
 
   it ('should return if a ParenthesisNode is associative with another Node', function () {
@@ -108,8 +106,8 @@ describe('operators', function () {
 
     var p = new ParenthesisNode(add);
 
-    assert.equal(operators.isAssociativeWith(p, sub, {parenthesis: 'all'}), true);
-    assert.equal(operators.isAssociativeWith(p, sub, {parenthesis: 'auto'}), true);
-    assert.equal(operators.isAssociativeWith(p, sub, config), null);
+    assert.equal(operators.isAssociativeWith(p, sub, 'all'), true);
+    assert.equal(operators.isAssociativeWith(p, sub, 'auto'), true);
+    assert.equal(operators.isAssociativeWith(p, sub, 'keep'), null);
   });
 });